Extract shared hooks and eliminate chart/image duplication
- useObjectKeyboardShortcuts: shared keyboard handler (Delete,
  Arrow, Escape) with external input guard
- useObjectDragResize: shared drag/resize effect with anchor
  re-calculation and draft management
- ObjectLayerViewport: shared viewport clipping wrapper
- DragState: unified type with objectId instead of chartId/imageId
- Add image staleness guard (clears selection when image is
  deleted remotely)

Reduces ~250 lines of duplicated code between chart-object-layer
and image-object-layer.
