+++ packages/frontend/src/app/spreadsheet/image-object-layer.tsx
@@ -4,16 +4,12 @@ import {
   useEffect,
   useReducer,
   useRef,
-  useState,
 } from "react";
 import { parseRef, type SheetImage, Spreadsheet } from "@wafflebase/sheets";
-import {
-  type DraftLayout,
-  type HandlePosition,
-  computeResizeDraft,
-  reanchorAfterMove,
-} from "./object-layer-utils";
+import { type DraftLayout, type HandlePosition } from "./object-layer-utils";
 import { SelectionOverlay } from "./selection-overlay";
+import { ObjectLayerViewport } from "./object-layer-viewport";
+import { useObjectKeyboardShortcuts, useObjectDragResize } from "./use-object-layer";
 import type { SpreadsheetDocument } from "@/types/worksheet";
 import { getOrLoadImage } from "./image-cache";
 
@@ -33,19 +29,6 @@ type ImageObjectLayerProps = {
   renderVersion: number;
 };
 
-type DragState = {
-  imageId: string;
-  mode: "move" | "resize";
-  handle?: HandlePosition;
-  startX: number;
-  startY: number;
-  startOffsetX: number;
-  startOffsetY: number;
-  startWidth: number;
-  startHeight: number;
-  aspectRatio: number;
-};
-
 // ---------------------------------------------------------------------------
 // ImageObjectLayer
 // ---------------------------------------------------------------------------
@@ -61,161 +44,27 @@ export function ImageObjectLayer({
   onDeleteImage,
   renderVersion,
 }: ImageObjectLayerProps) {
-  const [dragState, setDragState] = useState<DragState | null>(null);
-  const [drafts, setDrafts] = useState<Record<string, DraftLayout>>({});
-
   // Force re-render when images finish loading.
   const [, forceUpdate] = useReducer((x: number) => x + 1, 0);
 
   const images = Object.values(root.sheets[tabId]?.images || {});
-  const imagesRef = useRef(images);
-  imagesRef.current = images;
 
-  // Keyboard handler for image shortcuts (delete, move, escape).
-  useEffect(() => {
-    if (!selectedImageId || readOnly) return;
-    const onKeyDown = (event: KeyboardEvent) => {
-      // Don't intercept when user is typing in an external input (dialog, etc.).
-      // The grid's own cell input (contentEditable div) is NOT external.
-      const target = event.target;
-      if (target instanceof Element) {
-        const grid = document.querySelector("[data-sheet-container]");
-        if (!grid?.contains(target)) {
-          const tag = target.tagName;
-          if (
-            tag === "INPUT" ||
-            tag === "TEXTAREA" ||
-            (target as HTMLElement).isContentEditable
-          ) {
-            return;
-          }
-        }
-      }
+  useObjectKeyboardShortcuts({
+    selectedId: selectedImageId,
+    readOnly,
+    items: images,
+    onDelete: onDeleteImage,
+    onDeselect: () => onSelectImage(null),
+    onUpdate: onUpdateImage,
+  });
 
-      if (event.key === "Delete" || event.key === "Backspace") {
-        event.preventDefault();
-        event.stopPropagation();
-        onDeleteImage(selectedImageId);
-        onSelectImage(null);
-      } else if (
-        event.key === "ArrowUp" ||
-        event.key === "ArrowDown" ||
-        event.key === "ArrowLeft" ||
-        event.key === "ArrowRight"
-      ) {
-        event.preventDefault();
-        event.stopPropagation();
-        const dx =
-          event.key === "ArrowLeft" ? -10 : event.key === "ArrowRight" ? 10 : 0;
-        const dy =
-          event.key === "ArrowUp" ? -10 : event.key === "ArrowDown" ? 10 : 0;
-        onUpdateImage(selectedImageId, {
-          offsetX:
-            (images.find((i) => i.id === selectedImageId)?.offsetX ?? 0) + dx,
-          offsetY:
-            (images.find((i) => i.id === selectedImageId)?.offsetY ?? 0) + dy,
-        });
-      } else if (event.key === "Escape") {
-        event.stopPropagation();
-        onSelectImage(null);
-      }
-    };
-    // Use capture phase on document so this fires before the grid's
-    // keydown handler (which also listens on document in bubble phase).
-    document.addEventListener("keydown", onKeyDown, true);
-    return () => document.removeEventListener("keydown", onKeyDown, true);
-  }, [selectedImageId, readOnly, onDeleteImage, onSelectImage, onUpdateImage, images]);
-
-  // Drag/resize handler.
-  useEffect(() => {
-    if (!dragState || readOnly) return;
-
-    let latestX = dragState.startX;
-    let latestY = dragState.startY;
-
-    const toDraft = (clientX: number, clientY: number): DraftLayout => {
-      const z = spreadsheet?.getZoom() ?? 1;
-      const deltaX = (clientX - dragState.startX) / z;
-      const deltaY = (clientY - dragState.startY) / z;
-
-      if (dragState.mode === "move") {
-        return {
-          offsetX: dragState.startOffsetX + deltaX,
-          offsetY: dragState.startOffsetY + deltaY,
-          width: dragState.startWidth,
-          height: dragState.startHeight,
-        };
-      }
-
-      // Resize mode — images lock aspect ratio on corner handles.
-      return computeResizeDraft({
-        handle: dragState.handle!,
-        deltaX,
-        deltaY,
-        startOffsetX: dragState.startOffsetX,
-        startOffsetY: dragState.startOffsetY,
-        startWidth: dragState.startWidth,
-        startHeight: dragState.startHeight,
-        aspectRatio: dragState.aspectRatio,
-        lockAspectRatio: true,
-      });
-    };
-
-    const onPointerMove = (event: PointerEvent) => {
-      latestX = event.clientX;
-      latestY = event.clientY;
-      const nextDraft = toDraft(latestX, latestY);
-      setDrafts((prev) => ({
-        ...prev,
-        [dragState.imageId]: nextDraft,
-      }));
-    };
-
-    const onPointerUp = () => {
-      const nextDraft = toDraft(latestX, latestY);
-      const imageId = dragState.imageId;
-
-      // For move operations, re-anchor to the cell under the pointer
-      // so row/column insert/delete correctly shifts the image.
-      if (dragState.mode === "move" && spreadsheet) {
-        const img = imagesRef.current.find((i) => i.id === imageId);
-        if (img) {
-          const patch = reanchorAfterMove({
-            spreadsheet,
-            pointerX: latestX,
-            pointerY: latestY,
-            currentAnchor: img.anchor,
-            draft: nextDraft,
-          });
-          if (patch) {
-            onUpdateImage(imageId, patch);
-            setDrafts((prev) => {
-              const remaining = { ...prev };
-              delete remaining[imageId];
-              return remaining;
-            });
-            setDragState(null);
-            return;
-          }
-        }
-      }
-
-      onUpdateImage(imageId, nextDraft);
-      setDrafts((prev) => {
-        const remaining = { ...prev };
-        delete remaining[imageId];
-        return remaining;
-      });
-      setDragState(null);
-    };
-
-    window.addEventListener("pointermove", onPointerMove);
-    window.addEventListener("pointerup", onPointerUp);
-    return () => {
-      window.removeEventListener("pointermove", onPointerMove);
-      window.removeEventListener("pointerup", onPointerUp);
-    };
-  }, [dragState, onUpdateImage, readOnly, spreadsheet]);
+  const { setDragState, drafts } = useObjectDragResize({
+    readOnly,
+    spreadsheet,
+    lockAspectRatio: true,
+    items: images,
+    onUpdate: onUpdateImage,
+  });
 
   const mountedRef = useRef(false);
   useEffect(() => {
@@ -234,102 +83,66 @@ export function ImageObjectLayer({
   }
 
   const zoom = spreadsheet.getZoom();
-  const viewport = spreadsheet.getGridViewportRect();
-  const scrollableViewport = spreadsheet.getScrollableGridViewportRect();
-  const clipLeft = Math.max(0, scrollableViewport.left - viewport.left);
-  const clipTop = Math.max(0, scrollableViewport.top - viewport.top);
-  const clipWidth = Math.max(0, scrollableViewport.width);
-  const clipHeight = Math.max(0, scrollableViewport.height);
-
-  if (clipWidth === 0 || clipHeight === 0) {
-    return null;
-  }
 
   return (
-    <div
-      className="absolute pointer-events-none overflow-hidden"
-      data-render-version={renderVersion}
-      style={{
-        left: viewport.left,
-        top: viewport.top,
-        width: viewport.width,
-        height: viewport.height,
-        zIndex: 3,
-      }}
+    <ObjectLayerViewport
+      spreadsheet={spreadsheet}
+      zIndex={3}
+      renderVersion={renderVersion}
     >
-      <div
-        className="absolute pointer-events-none overflow-hidden"
-        style={{
-          left: clipLeft,
-          top: clipTop,
-          width: clipWidth,
-          height: clipHeight,
-        }}
-      >
-        <div
-          className="relative h-full w-full pointer-events-none"
-          style={{
-            left: -clipLeft,
-            top: -clipTop,
-            width: viewport.width,
-            height: viewport.height,
-          }}
-        >
-          {images.map((image) => {
-            const layout = drafts[image.id] || {
-              offsetX: image.offsetX,
-              offsetY: image.offsetY,
-              width: image.width,
-              height: image.height,
-            };
-            return (
-              <ImageObject
-                key={image.id}
-                image={image}
-                spreadsheet={spreadsheet}
-                zoom={zoom}
-                selected={selectedImageId === image.id}
-                readOnly={readOnly}
-                layout={layout}
-                onImageLoad={onImageLoad}
-                onSelect={() => onSelectImage(image.id)}
-                onMoveStart={(event) => {
-                  onSelectImage(image.id);
-                  setDragState({
-                    imageId: image.id,
-                    mode: "move",
-                    startX: event.clientX,
-                    startY: event.clientY,
-                    startOffsetX: layout.offsetX,
-                    startOffsetY: layout.offsetY,
-                    startWidth: layout.width,
-                    startHeight: layout.height,
-                    aspectRatio:
-                      layout.height > 0 ? layout.width / layout.height : 1,
-                  });
-                }}
-                onResizeStart={(event, handle) => {
-                  onSelectImage(image.id);
-                  setDragState({
-                    imageId: image.id,
-                    mode: "resize",
-                    handle,
-                    startX: event.clientX,
-                    startY: event.clientY,
-                    startOffsetX: layout.offsetX,
-                    startOffsetY: layout.offsetY,
-                    startWidth: layout.width,
-                    startHeight: layout.height,
-                    aspectRatio:
-                      layout.height > 0 ? layout.width / layout.height : 1,
-                  });
-                }}
-              />
-            );
-          })}
-        </div>
-      </div>
-    </div>
+      {images.map((image) => {
+        const layout = drafts[image.id] || {
+          offsetX: image.offsetX,
+          offsetY: image.offsetY,
+          width: image.width,
+          height: image.height,
+        };
+        return (
+          <ImageObject
+            key={image.id}
+            image={image}
+            spreadsheet={spreadsheet}
+            zoom={zoom}
+            selected={selectedImageId === image.id}
+            readOnly={readOnly}
+            layout={layout}
+            onImageLoad={onImageLoad}
+            onSelect={() => onSelectImage(image.id)}
+            onMoveStart={(event) => {
+              onSelectImage(image.id);
+              setDragState({
+                objectId: image.id,
+                mode: "move",
+                startX: event.clientX,
+                startY: event.clientY,
+                startOffsetX: layout.offsetX,
+                startOffsetY: layout.offsetY,
+                startWidth: layout.width,
+                startHeight: layout.height,
+                aspectRatio:
+                  layout.height > 0 ? layout.width / layout.height : 1,
+              });
+            }}
+            onResizeStart={(event, handle) => {
+              onSelectImage(image.id);
+              setDragState({
+                objectId: image.id,
+                mode: "resize",
+                handle,
+                startX: event.clientX,
+                startY: event.clientY,
+                startOffsetX: layout.offsetX,
+                startOffsetY: layout.offsetY,
+                startWidth: layout.width,
+                startHeight: layout.height,
+                aspectRatio:
+                  layout.height > 0 ? layout.width / layout.height : 1,
+              });
+            }}
+          />
+        );
+      })}
+    </ObjectLayerViewport>
   );
 }
 
@@ -433,4 +246,3 @@ function ImageObject({
     </div>
   );
 }
-
